+++ config/routes.rb
@@ -189,6 +189,7 @@ Rails.application.routes.draw do
       post "orders", to: "orders#create"
       post "payments/new_card", to: "payments#new_card"
       post "payments/one_click", to: "payments#one_click"
+      post "payments/sbp/init", to: "payments#sbp_init"
       get "payments/card_config", to: "payments#card_config"
       get "user/cards", to: "user_cards#index"
       get "orders/history", to: "orders#history"